+++ src/gui/minimap.cpp
@@ -56,8 +56,8 @@ Minimap::Minimap():
     setCloseButton(true);
 
     mShow = config.getValue(getWindowName() + "Show", true);
-    mUserWidth = config.getValue(getWindowName() + "UserWidth", 100);
-    mUserHeight = config.getValue(getWindowName() + "UserHeight", 100);
+    mUserWidth = (int) config.getValue(getWindowName() + "UserWidth", 100);
+    mUserHeight = (int) config.getValue(getWindowName() + "UserHeight", 100);
 
     setDefaultSize(100, 100, ImageRect::UPPER_LEFT, 5, 25);
